cosmetic: extract most of the outline-related functions into separate file
