+++ libass/Makefile.am
@@ -25,8 +25,8 @@ SRC_CORETEXT = ass_coretext.c ass_coretext.h
 
 lib_LTLIBRARIES = libass.la
 libass_la_SOURCES = ass.c ass_cache.c ass_font.c ass_fontselect.c ass_render.c \
-                    ass_utils.c ass_bitmap.c ass_blur.c ass_library.c ass_bitmap.h \
-                    ass_cache.h ass_fontselect.h ass_font.h ass.h \
+                    ass_utils.c ass_bitmap.c ass_outline.c ass_blur.c ass_library.c \
+                    ass_bitmap.h ass_outline.h ass_cache.h ass_fontselect.h ass_font.h ass.h \
                     ass_library.h ass_types.h ass_utils.h ass_drawing.c \
                     ass_drawing.h ass_cache_template.h ass_render.h \
                     ass_parse.c ass_parse.h ass_render_api.c ass_shaper.c \
