+++ libass/ass_font.h
@@ -76,13 +76,4 @@ FT_Glyph ass_font_get_glyph(ASS_Font *font,
                             ASS_Hinting hinting, int deco);
 void ass_font_clear(ASS_Font *font);
 
-void outline_translate(const ASS_Outline *outline, FT_Pos dx, FT_Pos dy);
-void outline_transform(const ASS_Outline *outline, const FT_Matrix *matrix);
-void outline_get_cbox(const ASS_Outline *outline, FT_BBox *cbox);
-void fix_freetype_stroker(ASS_Outline *outline, int border_x, int border_y);
-int outline_alloc(ASS_Outline *outline, size_t n_points, size_t n_contours);
-ASS_Outline *outline_convert(const FT_Outline *source);
-ASS_Outline *outline_copy(const ASS_Outline *source);
-void outline_free(ASS_Outline *outline);
-
 #endif                          /* LIBASS_FONT_H */
